Work on import errors and modify files
