+++ api/views/products_views.py
@@ -4,8 +4,6 @@ from flask_jwt_extended import jwt_required
 
 prodbp = Blueprint('prod', __name__)
 
-product_cls = Products()
-
 @prodbp.route('/products', methods=['POST'])
 # @jwt_required
 def add_product():
@@ -19,7 +17,8 @@ def add_product():
 
     if not product_name or not product_price or not product_stock:
         abort(400)
-
+    
+    product_cls = Products()
     product_cls.add_product(category, product_name,
                             product_specs, product_price, product_stock)
 
@@ -29,18 +28,21 @@ def add_product():
 @prodbp.route('/products', methods=['GET'])
 # @jwt_required
 def view_all_products():
+    product_cls = Products()
     return jsonify({"Products":product_cls.get_all_products()})
 
 
 @prodbp.route('/products/<int:product_id>', methods=['GET'])
 # @jwt_required
 def view_one_product(product_id):
+    product_cls = Products()
     return jsonify({"Product":product_cls.get_one_product_by_id(product_id)}), 200
 
 
 @prodbp.route('/products/<product_id>', methods=['DELETE'])
 # @jwt_required
 def delete_a_product(product_id):
+    product_cls = Products()
     product_cls.delete_a_product(product_id)
     return jsonify({"Deleted":"Product was deleted successfully"}), 200
 
@@ -51,6 +53,7 @@ def edit_product(product_id):
 
     product_stock = int(data['product_stock'])
     product_price = int(data['product_price'])
+    product_cls = Products()
     product_cls.edit_a_product(product_id, product_stock, product_price)
 
     return jsonify({"Updated":
