+++ requirements.txt
@@ -28,6 +28,7 @@ mccabe==0.6.1
 more-itertools==4.3.0
 pluggy==0.8.0
 psycopg2==2.7.5
+psycopg2-binary==2.7.5
 py==1.5.4
 pycodestyle==2.4.0
 pyflakes==2.0.0
